he creado la carpeta de los test y refactorizado el package de punto
