+++ code/src/main/java/com/endes/figuras/Punto.java
@@ -1,4 +1,4 @@
-package punto;
+package com.endes.figuras;
 
 public class Punto  {
 	double y;
