+++ code/src/test/java/com/endes/figuras/PuntoTest.java
@@ -1,0 +1,14 @@
+package com.endes.figuras;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.Test;
+
+class PuntoTest {
+
+	@Test
+	void test() {
+		fail("Not yet implemented");
+	}
+
+}
